refactor: move quality benchmark rendering into report module
Pure move: quality.rs had grown past 700 lines mixing orchestration,
parsing, and presentation; the TTY/JSON printing now lives in
benchmark/report.rs.
